Refactor id Integer to Long

+++ UnitTests/Java/DataJpaTest/src/main/java/com/restapi/demo/model/Product.java
@@ -1,9 +1,7 @@
 package com.restapi.demo.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 
 import javax.persistence.*;
 
@@ -15,7 +13,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 20)
     private String name;
